Switched to byte encoded messages

+++ controller/move.h
@@ -3,14 +3,14 @@
 
 #include <stdint.h>
 
-typedef uint8_t move;
+typedef uint16_t move;
 
 const move INVALID_MOVE = 0xFFFF;
 
 /**
  * @return an encoded move as origin-destination square
  */
-move encode(uint16_t from, uint16_t to) {
+move encode(uint8_t from, uint8_t to) {
     return (from << 8) | to;
 }
 /**
